refactor: fix linter errors
